refine docs
